Moving a couple classes to validation_utils.py module to stop a circular import.
